+++ kleidicv/src/resize/resize_linear_sc.h
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: 2024 - 2025 Arm Limited and/or its affiliates <open-source-office@arm.com>
+// SPDX-FileCopyrightText: 2024 - 2026 Arm Limited and/or its affiliates <open-source-office@arm.com>
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -12,7 +12,7 @@
 
 namespace KLEIDICV_TARGET_NAMESPACE {
 
-KLEIDICV_TARGET_FN_ATTRS static kleidicv_error_t resize_2x2_u8_sc(
+KLEIDICV_TARGET_FN_ATTRS static kleidicv_error_t kleidicv_resize_2x2_u8_sc(
     const uint8_t *src, size_t src_stride, size_t src_width, size_t src_height,
     size_t y_begin, size_t y_end, uint8_t *dst,
     size_t dst_stride) KLEIDICV_STREAMING {
@@ -157,7 +157,7 @@ KLEIDICV_TARGET_FN_ATTRS static kleidicv_error_t resize_2x2_u8_sc(
   return KLEIDICV_OK;
 }
 
-KLEIDICV_TARGET_FN_ATTRS static kleidicv_error_t resize_4x4_u8_sc(
+KLEIDICV_TARGET_FN_ATTRS static kleidicv_error_t kleidicv_resize_4x4_u8_sc(
     const uint8_t *src, size_t src_stride, size_t src_width, size_t src_height,
     size_t y_begin, size_t y_end, uint8_t *dst,
     size_t dst_stride) KLEIDICV_STREAMING {
@@ -1141,36 +1141,12 @@ KLEIDICV_TARGET_FN_ATTRS static kleidicv_error_t resize_8x8_f32_sve256plus_sc(
   return KLEIDICV_OK;
 }
 
-KLEIDICV_TARGET_FN_ATTRS static kleidicv_error_t resize_linear_stripe_u8_sc(
-    const uint8_t *src, size_t src_stride, size_t src_width, size_t src_height,
-    size_t y_begin, size_t y_end, uint8_t *dst, size_t dst_stride,
-    size_t dst_width, size_t dst_height) KLEIDICV_STREAMING {
-  CHECK_POINTER_AND_STRIDE(src, src_stride, src_height);
-  CHECK_POINTER_AND_STRIDE(dst, dst_stride, dst_height);
-  CHECK_IMAGE_SIZE(dst_width, dst_height);
-
-  if (src_width == 0 || src_height == 0) {
-    return KLEIDICV_OK;
-  }
-  if (src_width * 2 == dst_width && src_height * 2 == dst_height) {
-    return resize_2x2_u8_sc(src, src_stride, src_width, src_height, y_begin,
-                            y_end, dst, dst_stride);
-  }
-  if (src_width * 4 == dst_width && src_height * 4 == dst_height) {
-    return resize_4x4_u8_sc(src, src_stride, src_width, src_height, y_begin,
-                            y_end, dst, dst_stride);
-  }
-  // resize_linear_f32_is_implemented checked the kernel size already.
-  // GCOVR_EXCL_START
-  assert(!"resize ratio not implemented");
-  return KLEIDICV_ERROR_NOT_IMPLEMENTED;
-  // GCOVR_EXCL_STOP
-}
-
-KLEIDICV_TARGET_FN_ATTRS static kleidicv_error_t resize_linear_stripe_f32_sc(
-    const float *src, size_t src_stride, size_t src_width, size_t src_height,
-    size_t y_begin, size_t y_end, float *dst, size_t dst_stride,
-    size_t dst_width, size_t dst_height) KLEIDICV_STREAMING {
+KLEIDICV_TARGET_FN_ATTRS static kleidicv_error_t
+kleidicv_resize_linear_stripe_f32_sc(const float *src, size_t src_stride,
+                                     size_t src_width, size_t src_height,
+                                     size_t y_begin, size_t y_end, float *dst,
+                                     size_t dst_stride, size_t dst_width,
+                                     size_t dst_height) KLEIDICV_STREAMING {
   CHECK_POINTER_AND_STRIDE(src, src_stride, src_height);
   CHECK_POINTER_AND_STRIDE(dst, dst_stride, dst_height);
   CHECK_IMAGE_SIZE(dst_width, dst_height);